fix(client): code review comments
