+++ packages/amplication-client/src/User/UserQueries.ts
@@ -1,0 +1,12 @@
+import { gql } from "@apollo/client";
+
+export const SIGNUP_PREVIEW_ACCOUNT = gql`
+  mutation SignupPreviewAccount($data: SignupPreviewAccountInput!) {
+    signupPreviewAccount(data: $data) {
+      token
+      workspaceId
+      projectId
+      resourceId
+    }
+  }
+`;
